feat: move serializer to api common + add test for move operation
